master: Vagrantfile modification
I received an error when attempting to "vagrant reload" after the "vagrant up" completed.
Warning: Authentication failure. Retrying...
    
This error was caused by a problem with the ssh keys.  I modified the Vagrantfile to circumvent my issue based on what I found on:  https://github.com/hashicorp/vagrant/issues/5186
    
Everything seems to be fine (so far) with this small tweak.  I documented more information in the Vagrantfile comments above the statement I added.
    
This may not work for you.  For that matter, you may not be having this problem.  If this Vagrantfile works for you, more power to you.  I just wanted to help out someone else who might be having difficulty getting this vagrant script to work given bugs in Vagrant itself.

Changes to be committed:
    modified:   Vagrantfile